Added UserService test; moved contexts to resources; changed AbstractModel.id to Long (from int); externalized flyway config

+++ src/main/java/com/mskiba/archetype/dao/GenericDAO.java
@@ -4,7 +4,7 @@ import com.mskiba.archetype.model.AbstractModel;
 
 
 public interface GenericDAO<T extends AbstractModel> {
-    T getById(int id);
+    T getById(Long id);
 
     void save(T entity);
